pyproject.toml
[project]
name = "mcp-future-demo"
version = "0.1.0"
description = "End-to-end FastMCP demo covering core MCP primitives (tools, resources, prompts)"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "anthropic[mcp]>=0.39",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
]

[tool.uv]
package = false
